+++ src/main/kotlin/com/jetbrains/snakecharm/stringLanguage/lang/highlighter/SmkSLAnnotatingVisitor.kt
@@ -4,9 +4,10 @@ import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.openapi.project.DumbAware
 import com.intellij.psi.PsiElement
+import com.jetbrains.python.validation.PyAnnotationHolder
 
 class SmkSLAnnotatingVisitor : Annotator, DumbAware {
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
-        SmkSLWildcardsAnnotator.annotateElement(element, holder)
+        element.accept(SmkSLWildcardsAnnotator(PyAnnotationHolder(holder)))
     }
 }